refactor(command-visualizer): move render-model types to model/types.ts
Renames model/frame-state.ts -> model/types.ts and gives it a maintainer-
facing header: these container types (CascadeState/Frame/Decoration/FrameRole/
OverlayRole) are package-specific with no current cursorless home (field types
already come from lib-common). Isolated as a dedicated types file so cursorless
maintainers can decide whether any warrant promotion. Type-only rename +
import-path updates; zero behavior change (32 tests green, tsc clean).
