+++ packages/command-visualizer/src/logic/derive-overlays.ts
@@ -5,7 +5,7 @@
 // caller state. logic/ → logic/ import only.
 
 import type { OverlayStyleName } from "../data/decorations";
-import type { Decoration } from "../model/frame-state";
+import type { Decoration } from "../model/types";
 import {
   asArr,
   asObj,
